+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Optimizing `#.forEachEdge`.
 * Optimizing `#.mergeEdge`.
 * Optimizing `#.edges` and `#.nodes`.
+* Optimizing `#.import`.
 * Fixing `#.edges(A, B)` with directed self loops.
 * Fixing `#.mergeEdgeWithKey` edge cases.
 * Fixing `#.mergeAttributes` event payload.